c5978d7 Roll protocol to r623118 and update browser json patch

+++ chromewhip/protocol/indexeddb.py
@@ -18,7 +18,7 @@ from chromewhip.protocol import runtime as Runtime
 class DatabaseWithObjectStores(ChromeTypeBase):
     def __init__(self,
                  name: Union['str'],
-                 version: Union['int'],
+                 version: Union['float'],
                  objectStores: Union['[ObjectStore]'],
                  ):
